fix well-known dav redirect
nginx sets the scheme variable to the communication protocol it uses
with the reverse proxy. which is http, since the container doesn't
listen for https connections.

well-known urls therefore redirect to http://your-domain/remote.php/dav
instead of https://

this causes the browser to block the self check request in the admin
panel because of a content security policy that only allows self links
(and self is https:// and not http://)